add json names to all properties

+++ NetTools/Models/PingResult.cs
@@ -10,12 +10,12 @@ public record PingResult(
     [property: JsonPropertyName("data_bytes")] int DataBytes,
     [property: JsonPropertyName("packets_transmitted")] int PacketsTransmitted,
     [property: JsonPropertyName("packets_received")] int PacketsReceived,
-    [property: JsonPropertyName("packet_loss_percent")] float? PacketLossPercent,
-    [property: JsonPropertyName("time_ms")] float TotalTimeMs,
-    [property: JsonPropertyName("round_trip_time_min")] float RoundTripTimeMin,
-    [property: JsonPropertyName("round_trip_time_max")] float RoundTripTimeMax,
-    [property: JsonPropertyName("round_trip_time_avg")] float RoundTripTimeAvg,
-    [property: JsonPropertyName("round_trip_time_stdev")] float RoundTripTimeStddev,
+    [property: JsonPropertyName("packet_loss_percent")] double? PacketLossPercent,
+    [property: JsonPropertyName("time_ms")] double TotalTimeMs,
+    [property: JsonPropertyName("round_trip_time_min")] double RoundTripTimeMin,
+    [property: JsonPropertyName("round_trip_time_max")] double RoundTripTimeMax,
+    [property: JsonPropertyName("round_trip_time_avg")] double RoundTripTimeAvg,
+    [property: JsonPropertyName("round_trip_time_stdev")] double RoundTripTimeStddev,
     [property: JsonPropertyName("timestamp")] long? Timestamp,
     [property: JsonPropertyName("responses")] IReadOnlyList<PingResponse> Responses);
 
@@ -25,5 +25,5 @@ public record PingResponse(
     [property: JsonPropertyName("response_ip")] IPAddress ResponseIP,
     [property: JsonPropertyName("icmp_seq")] int IMCPSeq,
     [property: JsonPropertyName("ttl")] int TTL,
-    [property: JsonPropertyName("time_ms")] float TimeMs,
+    [property: JsonPropertyName("time_ms")] double TimeMs,
     [property: JsonPropertyName("duplicate")] bool Duplicate);
